chore(context-fabric): integrate CHAOS-3781 with 3786 and 3789 at rebase (CHAOS-3781)
Rebased onto main 9aece75. Three landed changes needed integration
beyond the mechanical migration renumber:

CHAOS-3786 (reuse identity chain) -- ReuseKey.ModelIdentity became
ModelIdentities []string with a membership test. My TimeAxisKey is a
separate, complementary dimension, so both survive: the key carries the
identity CHAIN and the axis, and FindReusable filters on
model_identity = ANY($5) AND time_axis_key = $7.

CHAOS-3789 (pull-request numbers as UInt32) -- overlapped my devhealthsource
scans, which added validity columns to the same SELECTs. Combined rather
than chosen: the scans keep 3789's uint32 typing AND carry the interval
columns.

CHAOS-3789's live schema-parity guard did its job and caught a real gap:
it builds a production-TYPED schema and runs every producer against it,
and my new columns were not declared there. Added with types read from
live ClickHouse system.columns rather than assumed -- work_items'
created_at is DateTime64(3) while git_pull_requests' is
DateTime64(3, 'UTC'), a distinction guessing would have gotten wrong. Its
seeds now name their columns, since a positional VALUES list no longer
lines up with the widened tables.

Migrations are contiguous again: 0012 is 3786's reuse-epoch cutover, 0013
is this branch's time-axis reuse key. expectedMigrationVersions is
{1..13} and the acr-migrate count is 13, both proved against real
Postgres.

Re-verified the DROP INDEX interaction against 3786's LANDED 0012: it
contains zero DDL statements and touches only rows in
context_fabric_investigation_results and context_fabric_reuse_invalidations.
My 0013 drops ix_acr_cf_investigation_results_reuse_key, created by 0011
and untouched by 0012. No interaction, confirmed on the merged code rather
than in advance.

+++ internal/contextfabric/devhealthsource/validity_test.go
@@ -100,7 +100,7 @@ func TestProducersEmitClosedValidityWindows(t *testing.T) {
 				"Investigate checkout flake", "done", "", at, created, uint8(1), ended}}
 		case "FROM git_pull_requests AS p":
 			// A merged pull request.
-			tables[index].rows = [][]any{{"repo-1", "example-org/widget-service", int64(1042),
+			tables[index].rows = [][]any{{"repo-1", "example-org/widget-service", uint32(1042),
 				"Typed session tokens", "merged", at, created, uint8(1), ended}}
 		case "FROM operational_incidents AS i":
 			// A resolved incident.
